migrate methods from board to game
